backend/requirements.txt
alembic==1.13.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
botocore==1.34.101
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.4
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.12.3
iso8601==2.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
limits==3.7.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
ndjson==0.3.1
ordered-set==4.1.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.4.11
platformdirs==4.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-mon==2.0.4
Pygments==2.17.2
pyhumps==3.8.0
PyJWT==2.8.0
pymon==1.1
pytest==7.2.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339==6.2
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.0
setuptools==69.1.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.0.7
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
WTForms==3.1.2
yarg==0.1.9
